change found_counter to unsigned long long
reason: no size_t printf specifier on some platforms

+++ src/main.c
@@ -16,7 +16,7 @@ pthread_mutex_t stdout_mtx = PTHREAD_MUTEX_INITIALIZER;
 #define safe_printf(x, ...) pthread_mutex_lock(&stdout_mtx); printf(x, ## __VA_ARGS__); \
                             pthread_mutex_unlock(&stdout_mtx)
 
-size_t found_counter = 0;
+unsigned long long found_counter = 0;
 pthread_mutex_t found_counter_mtx = PTHREAD_MUTEX_INITIALIZER;
 #define found_counter_increment() pthread_mutex_lock(&found_counter_mtx); ++found_counter; \
                                   pthread_mutex_unlock(&stdout_mtx)
@@ -120,7 +120,7 @@ int main(int argc, char *argv[])
 		}
 	}
 
-	safe_printf("%zu potential key(s) found!\n", found_counter);
+	safe_printf("%llu potential key(s) found!\n", found_counter);
 	fflush(stdout);
 	return EXIT_SUCCESS;
 }
